Fix: Problem on login and add login/out specs
By mistake we had 2 password fields on login, this removes unnecessary
double password field.

Also, to prevent similar cases in the future I've written specs for login and logout.
In the process I needed to change the way how we add password inside the
fixtures to be correctly hashed value.